+++ tests/Command/Mock/CommandMock.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Test\OctoLab\Cilex\Command\Mock;
+
+use OctoLab\Cilex\Command\Command;
+
+/**
+ * @author Kamil Samigullin <kamil@samigullin.info>
+ */
+class CommandMock extends Command
+{
+}