+++ tests/Command/Mock/HelloCommand.php
@@ -1,0 +1,31 @@
+<?php
+
+namespace Test\OctoLab\Cilex\Command\Mock;
+
+use OctoLab\Cilex\Command\Command;
+use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\Console\Input\InputOption;
+use Symfony\Component\Console\Output\OutputInterface;
+
+/**
+ * @author Kamil Samigullin <kamil@samigullin.info>
+ */
+class HelloCommand extends Command
+{
+    /**
+     * {@inheritdoc}
+     */
+    protected function configure()
+    {
+        $this->setName('hello')->addOption('message', 'm', InputOption::VALUE_REQUIRED, 'Welcome message');
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    protected function execute(InputInterface $input, OutputInterface $output)
+    {
+        $output->writeln(sprintf('Hello, %s', $input->getOption('message')));
+        return 0;
+    }
+}